Update folder structure Move state into one folder Remove unnecessary components
