--- answerstreams/src/app/features/comments/comments/comments.component.html
@@ -1,1 +1,0 @@
-<p>comments works!</p>
