--- answerstreams/src/app/features/comments/comments/comments.component.ts
@@ -1,13 +1,0 @@
-import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'app-comments',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './comments.component.html',
-  styleUrls: ['./comments.component.css']
-})
-export class CommentsComponent {
-
-}
