--- answerstreams/src/app/features/questions/questions.effects.ts
@@ -1,11 +1,0 @@
-import { Injectable } from "@angular/core"
-import { QuestionService } from "./questions.service"
-import { Actions, createEffect, ofType } from '@ngrx/effects'
-
-// @Injectable()
-// export const QuestionEffects {
-//     cosntructor(
-//         private actions$: Actions,
-//         private questionsService: QuestionService
-//     ){}
-// }
