--- answerstreams/src/app/features/users/users/users.component.html
@@ -1,1 +1,0 @@
-<p>users works!</p>
