--- answerstreams/src/app/features/users/users/users.component.ts
@@ -1,13 +1,0 @@
-import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'app-users',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './users.component.html',
-  styleUrls: ['./users.component.css']
-})
-export class UsersComponent {
-
-}
